+++ tests/legacy/media/show_poster_tests.py
@@ -18,7 +18,7 @@
 
 from __future__ import print_function
 
-from medusa.media.ShowPoster import ShowPoster
+from medusa.media.poster import ShowPoster
 from .generic_media_tests import GenericMediaTests
